Add files via upload
Working version. Changed algorithm to seek a minimum with coarse steps, interpolate the minimum double it and do a fine search there. Results are better for higher heart rates. In no minimum it goes to a coarse max search and interpolates a start position, then fine search. Changed main to use prox interrupt.